Refactor Section Member Model Location
