--- backend/models/courses/section_staff.py
@@ -1,16 +1,0 @@
-from pydantic import BaseModel
-
-# from backend.entities.courses.roster_role import RosterRole
-
-__authors__ = ["Ajay Gandecha"]
-__copyright__ = "Copyright 2023"
-__license__ = "MIT"
-
-
-class SectionStaff(BaseModel):
-    """
-    Pydantic model to represent staff members of course sections.
-    """
-
-    user_id: int
-    # role: RosterRole
