+++ backend/models/user_details.py
@@ -1,6 +1,5 @@
 from .permission import Permission
 from .user import User, UserIdentity
-from .roster_role import RosterRole
 
 __authors__ = ["Kris Jordan"]
 __copyright__ = "Copyright 2023"
@@ -29,18 +28,3 @@ class UserDetails(UserPermissions):
     """
 
     ...
-
-
-class SectionMember(UserIdentity):
-    """
-    Pydantic model to represent the information about a user who is a
-    staff of a section of a course.
-
-    This model is based on the `UserEntity` model, which defines the shape
-    of the `User` database in the PostgreSQL database
-    """
-
-    first_name: str
-    last_name: str
-    pronouns: str
-    member_type: RosterRole
